Fix warnings from clang v18.1.8

+++ DdsWicTest/dds.cpp
@@ -1352,14 +1352,14 @@ bool Test05(_In_ ID3D11Device* pDevice)
             if (!GetTempPathW(MAX_PATH, tempPath))
             {
                 success = false;
-                printf("ERROR: Getting temp path failed (%08X)\n", HRESULT_FROM_WIN32(GetLastError()));
+                printf("ERROR: Getting temp path failed (%08X)\n", static_cast<unsigned int>(HRESULT_FROM_WIN32(GetLastError())));
                 continue;
             }
 
             if (!GetTempFileNameW(tempPath, L"screenGrab", 0, tempFileName))
             {
                 success = false;
-                printf("ERROR: Getting temp file failed (%08X)\n", HRESULT_FROM_WIN32(GetLastError()));
+                printf("ERROR: Getting temp file failed (%08X)\n", static_cast<unsigned int>(HRESULT_FROM_WIN32(GetLastError())));
                 continue;
             }
